+++ src/core/program.py
@@ -20,15 +20,15 @@
 # along with Pashmak.  If not, see <https://www.gnu.org/licenses/>.
 #########################################################################
 
-""" Pashmak program object """
+""" Pashmak program Exceutor """
 
 import sys
 import os
 import signal
 from pathlib import Path
-from core import parser
-from core import helpers, version, modules, jit
+from core import helpers, version, modules, jit, parser
 from core.class_system import Class
+from core.function import Function
 
 import hashlib, time, random, datetime
 
@@ -36,17 +36,18 @@ class Program(helpers.Helpers):
     """ Pashmak program object """
 
     def __init__(self, is_test=False, args=[]):
-        self.states = [{
+        self.threads = [{
             'current_step': 0,
-            'commands': [],
+            'commands': [parser.parse('pass')[0]],
+            'used_namespaces': [],
             'vars': {
                 'argv': args,
                 'argc': len(args)
             }
-        }] # list of states
+        }] # list of threads
         self.functions = {
-            "mem": [], # mem is a empty function just for save mem in code
-            "rmem": [],
+            "mem": Function(name='mem', prog=self), # mem is a empty function just for save mem in code
+            "rmem": Function(name='rmem', prog=self),
         } # declared functions <function-name>:[<list-of-body-commands>]
         self.sections = {} # list of declared sections <section-name>:<index-of-command-to-jump>
         self.classes = {} # list of declared classes
@@ -57,12 +58,11 @@ class Program(helpers.Helpers):
         self.runtime_error = None # program raised error (for testing state)
         self.try_endtry = [] # says program is in try-endtry block
         self.namespaces_tree = [] # namespaces tree
-        self.used_namespaces = [] # list of used namespaces
         self.included_modules = [] # list of included modules to stop repeating imports
 
         self.allowed_pashmak_extensions = ['pashm']
 
-        self.states[-1]['current_step'] = 0
+        self.threads[-1]['current_step'] = 0
         self.stop_after_error = True
         self.main_filename = os.getcwd() + '/__main__'
 
@@ -72,7 +72,7 @@ class Program(helpers.Helpers):
 
     def import_script(self, paths, import_once=False):
         """ Imports scripts/modules """
-        op = self.states[-1]['commands'][self.states[-1]['current_step']]
+        op = self.threads[-1]['commands'][self.threads[-1]['current_step']]
 
         if type(paths) == str:
             paths = [paths]
@@ -117,24 +117,17 @@ class Program(helpers.Helpers):
 
     def set_commands(self, commands: list):
         """ Set commands list """
-        # include stdlib before everything
-        tmp = parser.parse('''
-        $__file__ = "''' + os.path.abspath(self.main_filename).replace('\\', '\\\\') + '''"
-        $__dir__ = "''' + os.path.dirname(os.path.abspath(self.main_filename)).replace('\\', '\\\\') + '''"
-        $__ismain__ = True
-        mem self.import_script('@stdlib')
-        ''', filepath='<system>')
-        commands.insert(0, tmp[0])
-        commands.insert(1, tmp[1])
-        commands.insert(2, tmp[2])
-        commands.insert(3, tmp[3])
-
+        # setup environment
+        self.set_var('__file__', os.path.abspath(self.main_filename))
+        self.set_var('__dir__', os.path.dirname(os.path.abspath(self.main_filename)))
+        self.set_var('__ismain__', True)
+        self.import_script('@stdlib')
         # set commands on program object
-        self.states[-1]['commands'] = commands
+        self.threads[-1]['commands'] = commands
 
     def set_command_index(self, op: dict) -> dict:
         """ Add command index to command dictonary """
-        op['index'] = self.states[-1]['current_step']
+        op['index'] = self.threads[-1]['current_step']
         return op
 
     def get_mem(self):
@@ -143,17 +136,6 @@ class Program(helpers.Helpers):
         self.mem = None
         return mem
 
-    def update_section_indexes(self, after_index):
-        """
-        When a new command inserted in commands list,
-        this function add 1 to section indexes to be
-        sync with new commands list
-        """
-        for k in self.sections:
-            if self.sections[k] > after_index:
-                self.sections[k] = self.sections[k] + 1
-        i = 0
-
     def raise_error(self, error_type: str, message: str, op: dict):
         """ Raise error in program """
         # check is in try
@@ -161,7 +143,7 @@ class Program(helpers.Helpers):
             section_index = self.try_endtry[-1]
             self.try_endtry.pop()
             new_step = self.sections[str(section_index)]
-            self.states[-1]['current_step'] = new_step-1
+            self.threads[-1]['current_step'] = new_step-1
 
             # put error data in mem
             self.mem = {'type': error_type, 'message': message, 'index': op['index']}
@@ -170,57 +152,61 @@ class Program(helpers.Helpers):
         if self.is_test:
             self.runtime_error = {'type': error_type, 'message': message, 'index': op['index']}
             if self.stop_after_error:
-                self.states[-1]['current_step'] = len(self.states[-1]['commands'])*2
+                self.threads[-1]['current_step'] = len(self.threads[-1]['commands'])*2
             return
 
         # render error
         print(error_type + ': ' + message + ':')
-        last_state = self.states[0]
-        for state in self.states[1:]:
+        last_thread = self.threads[0]
+        for thread in self.threads[1:]:
             try:
-                if last_state:
-                    tmp_op = last_state['commands'][last_state['current_step']]
+                if last_thread:
+                    tmp_op = last_thread['commands'][last_thread['current_step']]
                 else:
-                    tmp_op = state['commands'][0]
+                    tmp_op = thread['commands'][0]
                 print(
                     '\tin ' + tmp_op['file_path'] + ':' + str(tmp_op['line_number'])\
                     + ': ' + tmp_op['str']
                 )
             except KeyError:
                 pass
-            last_state = state
+            last_thread = thread
         print('\tin ' + op['file_path'] + ':' + str(op['line_number']) + ': ' + op['str'])
         sys.exit(1)
 
-    def exec_func(self, func_body: list, with_state=True, default_variables={}):
+    def exec_func(self, func_body: list, with_thread=True, default_variables={}):
         """ Gets a list from commands and runs them as function or included script """
         old_dir = self.get_var('__dir__')
         old_file = self.get_var('__file__')
-        # create new state for this call
-        if with_state:
-            state_vars = dict(self.states[-1]['vars'])
+        # create new thread for this call
+        if with_thread:
+            thread_vars = dict(self.threads[-1]['vars'])
         else:
-            state_vars = self.states[-1]['vars']
+            thread_vars = self.threads[-1]['vars']
 
         for k in default_variables:
-            state_vars[k] = default_variables[k]
+            thread_vars[k] = default_variables[k]
         if len(func_body) > 0:
-            state_vars['__file__'] = func_body[0]['file_path']
-            if os.path.isfile(state_vars['__file__']):
-                state_vars['__dir__'] = os.path.dirname(state_vars['__file__'])
-        self.states.append({
+            thread_vars['__file__'] = func_body[0]['file_path']
+            if os.path.isfile(thread_vars['__file__']):
+                thread_vars['__dir__'] = os.path.dirname(thread_vars['__file__'])
+        used_namespaces = []
+        if not with_thread:
+            used_namespaces = self.threads[-1]['used_namespaces']
+        self.threads.append({
             'current_step': 0,
             'commands': func_body,
-            'vars': state_vars,
+            'vars': thread_vars,
+            'used_namespaces': used_namespaces,
         })
 
         # run function
-        self.start_state()
+        self.start_thread()
 
         self.set_var('__dir__', old_dir)
         self.set_var('__file__', old_file)
 
-    def eval(self, command, only_parse=False):
+    def eval(self, command, only_parse=False, varname_as_dict=False, only_str_parse=False, dont_check_vars=False):
         """ Runs eval on command """
         i = 0
         command = command.strip()
@@ -257,6 +243,9 @@ class Program(helpers.Helpers):
                 command_parts[-1][1] += command[i]
             i += 1
 
+        if only_str_parse:
+            return command_parts
+
         full_op = ''
         opened_inline_calls_count = 0
         for code in command_parts:
@@ -271,8 +260,12 @@ class Program(helpers.Helpers):
                         if word[0] == '$' and not '@' in word:
                             variables_in_code.append(word[1:])
                 for k in variables_in_code:
-                    self.variable_required(k, self.states[-1]['commands'][self.states[-1]['current_step']])
-                    code = code.replace('$' + k, 'self.get_var("' + k + '")')
+                    if dont_check_vars == False:
+                        self.variable_required(k, self.threads[-1]['commands'][self.threads[-1]['current_step']])
+                    if varname_as_dict:
+                        code = code.replace('$' + k, 'self.all_vars()["' + k + '"]')
+                    else:
+                        code = code.replace('$' + k, 'self.get_var("' + k + '")')
                 code = code.replace('->', '.')
                 code = code.replace('^', 'self.get_mem()')
                 z = 0
@@ -308,7 +301,7 @@ class Program(helpers.Helpers):
         return eval(full_op)
 
     def call_inline_func(self, code: str):
-        """ Runs the internal function call "%{ func_or_command }" """
+        """ Runs the internal function call "%{func_or_command }" """
         commands = parser.parse(code)
         self.exec_func(commands, False)
         return self.get_mem()
@@ -332,9 +325,9 @@ class Program(helpers.Helpers):
             self.current_func
             try:
                 self.current_class
-                self.classes[self.current_class].methods[self.current_func].append(op)
+                self.classes[self.current_class].methods[self.current_func].body.append(op)
             except:
-                self.functions[self.current_func].append(op)
+                self.functions[self.current_func].body.append(op)
             return
         except NameError:
             pass
@@ -380,14 +373,7 @@ class Program(helpers.Helpers):
                 op_func(op)
             return
 
-        # check command syntax is variable value setting
-        tmp_bool = True
         if op['str'][0] == '$':
-            tmp_parts = op['str'].strip().split('@', 1)
-            if self.variable_exists(tmp_parts[0].strip()[1:]) and len(tmp_parts) > 1:
-                tmp_bool = False
-
-        if op['str'][0] == '$' and tmp_bool:
             # if a class is started, append current command as a property to class
             is_in_class = False
             try:
@@ -399,7 +385,11 @@ class Program(helpers.Helpers):
                 pass
             except AttributeError:
                 pass
-            parts = op['str'].strip().split('=', 1)
+            parts = self.split_by_equals(op['str'].strip()) # op['str'].strip().split('=', 1)
+            if len(parts) <= 1:
+                if '->' in op['str']:
+                    self.mem = self.eval(op['str'])
+                    return
             varname = parts[0].strip()
             full_varname = varname
             varname = varname.split('->', 1)
@@ -431,52 +421,36 @@ class Program(helpers.Helpers):
             return
 
         # check function exists
-        is_method = False
-        if op['command'][0] == '$':
-            var_name = op['command'].split('@')[0]
-            var = self.all_vars()[var_name[1:]]
-            if type(var) != Class:
-                return self.raise_error('MethodError', 'calling method on non-class object "' + var_name + '"', op)
-            try:
-                func_body = var.methods[op['command'].split('@')[1]]
-                is_method = var
-            except:
-                return self.raise_error('MethodError', 'class ' + self.all_vars()[var_name[1:]].__name__ + ' has not method "' + op['command'][0].split('@')[1] + '"', op)
-        else:
-            try:
-                func_body = self.functions[self.current_namespace() + op_name]
-            except KeyError:
-                func_body = None
-                for used_namespace in self.used_namespaces:
-                    try:
-                        func_body = self.functions[used_namespace + '.' + op_name]
-                    except KeyError:
-                        pass
-                if not func_body:
-                    try:
-                        func_body = self.functions[op_name]
-                    except KeyError:
-                        return self.raise_error('SyntaxError', 'undefined function "' + op_name + '"', op)
+        try:
+            func_body = self.functions[self.current_namespace() + op_name]
+        except KeyError:
+            func_body = None
+            for used_namespace in self.threads[-1]['used_namespaces']:
+                try:
+                    func_body = self.functions[used_namespace + '.' + op_name]
+                except KeyError:
+                    pass
+            if not func_body:
+                try:
+                    func_body = self.functions[op_name]
+                except KeyError:
+                    return self.raise_error('SyntaxError', 'undefined function "' + op_name + '"', op)
 
         # run function
         try:
             # put argument in the mem
-            if op['args_str'] != '':
+            if op['args_str'] != '' and op['args_str'].strip() != '()':
                 if op['command'] != 'rmem':
                     self.mem = self.eval(op['args_str'])
                 else:
                     self.eval(op['args_str'])
-            else:
-                self.mem = ''
 
             # execute function body
-            with_state = True
+            with_thread = True
             if op_name in ['import', 'mem', 'python', 'rmem', 'eval']:
-                with_state = False
+                with_thread = False
             default_variables = {}
-            if is_method != False:
-                default_variables['this'] = is_method
-            self.exec_func(func_body, with_state, default_variables=default_variables)
+            self.exec_func(func_body.body, with_thread, default_variables=default_variables)
             return
         except Exception as ex:
             raise
@@ -528,44 +502,44 @@ class Program(helpers.Helpers):
                                 raise
                                 pass
 
-    def start_state(self):
-        """ Start running state thread """
+    def start_thread(self):
+        """ Start running last thread """
         is_in_func = False
-        self.states[-1]['current_step'] = 0
+        self.threads[-1]['current_step'] = 0
 
         # load the sections
         i = 0
-        while i < len(self.states[-1]['commands']):
-            current_op = self.set_command_index(self.states[-1]['commands'][i])
+        while i < len(self.threads[-1]['commands']):
+            current_op = self.set_command_index(self.threads[-1]['commands'][i])
             if current_op['command'] == 'section':
                 if not is_in_func:
                     arg = current_op['args'][0]
                     self.sections[arg] = i+1
-                    self.states[-1]['commands'][i] = parser.parse('pass', filepath='<system>')[0]
+                    self.threads[-1]['commands'][i] = parser.parse('pass', filepath='<system>')[0]
             elif current_op['command'] == 'func':
                 is_in_func = True
             elif current_op['command'] == 'endfunc':
                 is_in_func = False
             i += 1
 
-        self.states[-1]['current_step'] = 0
-        while self.states[-1]['current_step'] < len(self.states[-1]['commands']):
+        self.threads[-1]['current_step'] = 0
+        while self.threads[-1]['current_step'] < len(self.threads[-1]['commands']):
             try:
-                self.run(self.states[-1]['commands'][self.states[-1]['current_step']])
+                self.run(self.threads[-1]['commands'][self.threads[-1]['current_step']])
             except Exception as ex:
                 try:
-                    self.set_command_index(self.states[-1]['commands'][self.states[-1]['current_step']])
+                    self.set_command_index(self.threads[-1]['commands'][self.threads[-1]['current_step']])
                 except:
                     break
                 self.raise_error(
                     ex.__class__.__name__,
                     str(ex),
-                    self.set_command_index(self.states[-1]['commands'][self.states[-1]['current_step']])
+                    self.set_command_index(self.threads[-1]['commands'][self.threads[-1]['current_step']])
                 )
-            self.states[-1]['current_step'] += 1
+            self.threads[-1]['current_step'] += 1
 
-        if len(self.states) > 1:
-            self.states.pop()
+        if len(self.threads) > 1:
+            self.threads.pop()
 
     def start(self):
         """ Start running the program """
@@ -574,4 +548,4 @@ class Program(helpers.Helpers):
 
         self.bootstrap_modules()
 
-        self.start_state()
+        self.start_thread()