+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Pashmak Changelog
 
-## 0.6, 0.6-beta5
+## 0.6, 0.6-beta6
 
 #### New features
 - added **Class** system
@@ -11,7 +11,7 @@
 - added `return` command for functions
 - added `import_once` function
 - added a syntax to handle dictonary item value assigning, now we can run `$my_dict["item"] = "value"`
-- added inline function calling feature, for example `println %{ my_func_or_command %{ another_command_as_arg }% }%`
+- added inline function calling feature, for example `println %{my_func_or_command %{another_command_as_arg}%}%`
 - added general variable `$__ismain__`
 - added **Jit** compiler
 
@@ -22,11 +22,15 @@
 - fixed lot of interpreter internal bugs and some optimizations
 - fixed a bug in function declaration syntax
 - fixed a bug in `$__dir__` and `$__file__` variables in function
+- fixed a bug in multi importing syntax
+- fixed some bugs in `end*` commands
+- fixed import command bug while using `()`
 
 #### Removed
 - deleted `loop` and `while` functions
 - deleted `out`, `fread`, `include` and `fwrite` commands because they was very low level and are not needed
-- deleted `$var = ^ command` syntax(use `$var = %{ command }%` instead of that)
+- deleted `$var = ^ command` syntax(use `$var = %{command}%` instead of that)
+- deleted `@file` module(use python file api instead of that)
 
 #### Other changes
 - change default `RuntimeError` name to the real python exception name in exception system